+++ internal/cmd/files/find/find.go
@@ -11,7 +11,7 @@ import (
 	"github.com/spf13/viper"
 )
 
-var exclusions = file.DefaultExclusions()
+var exclusions = file.Exclusions()
 var jsonPrint bool
 var lockfileOnly bool
 var strictness int
@@ -45,6 +45,8 @@ Special Terms | Meaning
 "?"           | matches any single non-Separator character
 "[class]"     | matches any single non-Separator character against a class of characters ([see "character classes"])
 "{alt1,...}"  | matches a sequence of characters if one of the comma-separated alternatives matches
+
+Exclude flags could alternatively be set using DEBRICKED_EXCLUSIONS="path1,path2,path3".
 
 Example: 
 $ debricked files find . `+exampleFlags)
