+++ internal/file/exclusion.go
@@ -9,19 +9,23 @@ import (
 const debrickedExclusionEnvVar = "DEBRICKED_EXCLUSIONS"
 
 func DefaultExclusions() []string {
-	defaultValues := []string{
+	return []string{
 		filepath.Join("**", "node_modules", "**"),
 		filepath.Join("**", "vendor", "**"),
 		filepath.Join("**", ".git", "**"),
 		filepath.Join("**", "obj", "**"), // nuget
 	}
+}
+
+func Exclusions() []string {
+	values := DefaultExclusions()
 
 	envValue := os.Getenv(debrickedExclusionEnvVar)
 	if envValue != "" {
-		defaultValues = strings.Split(envValue, ",")
+		values = strings.Split(envValue, ",")
 	}
 
-	return defaultValues
+	return values
 }
 
 var EXCLUDED_DIRS_FINGERPRINT = []string{
